BTWCT-12866-ui-irequest-create-loading-image-on-support-case-submission-page Modified the code so that the common loading image is utilised for the waiting periods for the two Mule API Calls.
Using method used in eProc. This made the changes and the ease of implementation a lot simpler. This method utilises a div that's rendered in the header component for Eproc only. As a result, in the header section, the switch to only render for Eproc has been removed.

<div data-sly-test="${header.eprocurement}" id="cmp-header--loader"></div>   **  has changed to

<div id="cmp-header--loader"></div>

Small changes to the create-request-form.js & the submit.js logic to display and hide the modal at the correct times
Change to not redirect in Edit mode when the Global Node is configured to redirect

Small changes to the CSS for the Back link