BTWCT-12886-stage-irequest-italy-screen-junk-values-are-displaying Modified the text-with-links.js code so that it uses the htmlParser function from the eCommerce functions to render the text and the link text. Also had to alter the AEM Parameter context="text" to context="unsafe".
Modified the form.js code so that it uses the htmlParser function from the eCommerce functions to modify the submit button text. Also had to alter the AEM Parameter context="text" to context="unsafe".

Note
In theory, the translation process could result in any of the labels having apostrophe's in so I've changed all the context='text' to context='unsafe' for all labels and text rendered for the forms at design time.